further organize dirs, add smoke tests and github workflow
