+++ .github/workflows/smoke_test.yml
@@ -1,0 +1,30 @@
+on:
+  push:
+    branches:
+      - main
+      - feature/**
+      - release/**'
+    pr:
+      - main
+      - next
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node: [ '12', '14', '16' ]
+    name: Node ${{ matrix.node }} sample
+    steps:
+      - uses: actions/checkout@v2
+      - name: Setup node
+        uses: actions/setup-node@v2
+        with:
+          node-version: ${{ matrix.node }}
+      - run: yarn
+      - name: 'Run tests'
+        run: yarn jest
+      - name: 'Build library'
+        run: yarn build
+
+
+
