+++ package.json
@@ -15,16 +15,21 @@
     "react-dom": "^17.0.2"
   },
   "devDependencies": {
+    "@testing-library/react": "^12.1.3",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.4.0",
     "@types/react": "^17.0.33",
     "@types/react-dom": "^17.0.10",
     "@vitejs/plugin-react": "^1.0.7",
     "autoprefixer": "^10.4.2",
+    "jest": "^27.5.1",
     "postcss": "^8.4.6",
     "postcss-apply": "^0.12.0",
     "postcss-extend": "^1.0.5",
     "postcss-import": "^14.0.2",
     "postcss-nesting": "^10.1.2",
     "tailwindcss": "^3.0.23",
+    "ts-jest": "^27.1.3",
     "typescript": "^4.5.4",
     "vite": "^2.8.0"
   }
