+++ tailwind.config.js
@@ -3,6 +3,7 @@ const colors = require('tailwindcss/colors')
 module.exports = {
   content: [
     "./index.html",
+    "./src/*.{vue,js,ts,jsx,tsx}",
     "./src/**/*.{vue,js,ts,jsx,tsx}",
     "./src/styles/**/*"
   ],
